feat: 랜딩과 로그인 페이지에서 사용되는 header를 auth-layout으로 분리
